requirements.txt
Flask==3.0.0
Flask_MySQLdb==2.0.0
PyJWT==2.8.0
